♻️ F03: Refactor Architecture, Background & Logo
