+++ src/components/pages/login/LoginSeparate.tsx
@@ -2,7 +2,7 @@ export default function Separate() {
   const hrStyle = {
     width: "100%",
     borderTop: "none",
-    borderBottom: `1px solid #F56A2C`,
+    borderBottom: `2px solid #F56A2C`,
     margin: "14.16px 0 22px",
   };
